@@ -1,22 +1,31 @@
 package org.jackstaff.grpc.demo;
 
 import org.jackstaff.grpc.annotation.AsynchronousUnary;
+import org.jackstaff.grpc.annotation.UnaryServerStreaming;
 
+import java.util.List;
 import java.util.function.Consumer;
 
 public interface AdvancedHelloService {
 
-    @AsynchronousUnary
-    void postMessage(String message); //Asynchronous Unary RPCs
+    int postMessage(String message); //Asynchronous Unary RPCs
 
     @AsynchronousUnary
-    void postMessage(String message, Consumer<String> result); //Asynchronous Unary RPCs
+    default void postMessage(String message, Consumer<Integer> result){ //Asynchronous Unary RPCs
+        result.accept(postMessage(message));
+    }
 
     String deny(String message);
 
     String sayHello(String greeting); //Unary RPCs
 
     void lotsOfReplies(String greeting, Consumer<HelloResponse> replies);//Server streaming RPCs
+
+    @UnaryServerStreaming
+    default List<HelloResponse> lotsOfReplies(String greeting){
+        System.out.println("NEVER............");
+        return null;
+    }
 
     Consumer<HelloRequest> lotsOfGreetings(SocialInfo socialInfo); //Client streaming RPCs
 
