@@ -14,7 +14,7 @@ public class ClientConfig {
     private int maxInboundMessageSize;
     private int idleTimeout;
     private boolean plaintext=true;
-    private int defaultTimeout=-1;
+    private int defaultUnaryTimeout;
 
     public String getHost() {
         return host;
@@ -88,12 +88,12 @@ public class ClientConfig {
         this.plaintext = plaintext;
     }
 
-    public int getDefaultTimeout() {
-        return defaultTimeout;
+    public int getDefaultUnaryTimeout() {
+        return defaultUnaryTimeout;
     }
 
-    public void setDefaultTimeout(int defaultTimeout) {
-        this.defaultTimeout = defaultTimeout;
+    public void setDefaultUnaryTimeout(int defaultUnaryTimeout) {
+        this.defaultUnaryTimeout = defaultUnaryTimeout;
     }
 
 }
